Fixed major error

+++ hw1/src/edu/upenn/cis573/hwk1/output/Output.java
@@ -12,5 +12,5 @@ public interface Output {
 
 	public void displayFileResults(File file, DecryptionResult decryptionResult);
 	
-	public void displayAggregateResults(int correct, int incorrect);
+	public void displayAggregateResults(long correct, long incorrect);
 }
